% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R, R/remove.R, R/replace.R
\name{edit-file}
\alias{edit-file}
\alias{edit_insert}
\alias{edit_remove}
\alias{edit_replace}
\title{Edit file}
\usage{
edit_insert(path, code, before = NULL, after = NULL, style = TRUE)

edit_remove(path, selection, style = TRUE)

edit_replace(path, code, selection, style = TRUE)
}
\arguments{
\item{path}{String. Path to file to edit}

\item{code}{Character vector. Code to insert or use as a replacement. If a language
object (type "language" or "symbol") it will be deparsed, if a function it will be applied on the matched
code, if a formula it will be converted to a function with \code{rlang::as_function()} and
applied on the matched code.}

\item{before, after}{<\code{\link[tidyselect:language]{tidy-select}}> using \link[=editor_tidy_select]{editor helpers}. destination of \code{code}, use either, not both and not not none.}

\item{style}{Boolean, wether to style the output. Useful because it's hard to monitor
indention, so set to \code{TRUE} by default. It also conveniently fails when the produced
code is not syntactic, before editing the file.}

\item{selection}{<\code{\link[tidyselect:language]{tidy-select}}> using \link[=editor_tidy_select]{editor helpers}. Location of code to replace or remove.}
}
\value{
Returns \code{NULL} invisibly, called for side effects.
}
\description{
Edit file
}
