% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{edit_insert}
\alias{edit_insert}
\title{Insert code in file}
\usage{
edit_insert(path, code, before = NULL, after = NULL, style = TRUE)
}
\arguments{
\item{path}{String. Path to file to edit}

\item{code}{Character vector. Code to insert}

\item{before, after}{\if{html}{\out{<tidy-select>}} using {editor} helpers. destination of \code{code}, use either, not both and not not none.}

\item{style}{Boolean, wether to style the output. Useful because it's hard to monitor
indention, so set to \code{TRUE} by default}
}
\description{
Insert code in file
}
