% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{edit_remove}
\alias{edit_remove}
\title{Remove code from file}
\usage{
edit_remove(path, selection, style = TRUE)
}
\arguments{
\item{path}{String. Path to file to edit}

\item{style}{Boolean, wether to style the output. Useful because it's hard to monitor
indention, so set to \code{TRUE} by default}

\item{code}{Character vector. Code to insert}

\item{before, after}{\if{html}{\out{<tidy-select>}} using {editor} helpers. destination of \code{code}, use either, not both and not not none.}
}
\description{
Remove code from file
}
