% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{edit_replace}
\alias{edit_replace}
\title{Insert code in file}
\usage{
edit_replace(path, code, selection, style = TRUE)
}
\arguments{
\item{path}{String. Path to file to edit}

\item{code}{Character vector. Code to insert}

\item{style}{Boolean, wether to style the output. Useful because it's hard to monitor
indention, so set to \code{TRUE} by default}

\item{before, after}{\if{html}{\out{<tidy-select>}} using {editor} helpers. destination of \code{code}, use either, not both and not not none.}
}
\description{
Insert code in file
}
