% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor-package.R
\docType{package}
\name{editor-package}
\alias{editor}
\alias{editor-package}
\title{editor: Edit scripts programatically}
\description{
A set of functions t easily insert, replace or remove code in/from a script.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/editor}
  \item Report bugs at \url{https://github.com/moodymudskipper/editor/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
