% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{editorUI}
\alias{editor}
\alias{editorUI}
\title{Editor Module}
\usage{
editorUI(id)

editor(input, output, session, path)
}
\arguments{
\item{id}{id to identify the UI}

\item{input}{for Shiny Module}

\item{output}{for Shiny Module}

\item{session}{for Shiny Module}

\item{path}{the path for a csv file}
}
\description{
UI and Server for editor Shiny Module
}

