% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect-helpers.R
\name{expr_calls}
\alias{expr_calls}
\title{Location of matching call}
\usage{
expr_calls(fun, n = NULL)
}
\arguments{
\item{var}{function name to match, or pattern such as returned by \code{stringr::regex()}
or other modifiers.}
}
\description{
Location of matching call
}
