% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect-helpers.R
\name{expr_defines}
\alias{expr_defines}
\title{Location of matching variable definitions}
\usage{
expr_defines(var, n = NULL)
}
\arguments{
\item{var}{variable name to match, or pattern such as returned by \code{stringr::regex()}
or other modifiers.}
}
\description{
Location of matching variable definitions
}
