% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect-helpers.R
\name{in_function}
\alias{in_function}
\title{Locate code inside a function definition}
\usage{
in_function(fun, selection, n = NULL)
}
\arguments{
\item{fun, n}{passed to \code{expr_defines()} to locate a function definition}

\item{selection}{selection to apply inside the function's body}
}
\description{
Locate code inside a function definition
}
