% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect-helpers.R
\name{line_matches}
\alias{line_matches}
\title{Location of matching code}
\usage{
line_matches(code, n = NULL)
}
\arguments{
\item{code}{code to match, or pattern such as returned by \code{stringr::regex()}
or other modifiers.}
}
\description{
Find matching line, by default we look for perfectly matching lines after trimming
the leading and trailing white spaces, stringr modifiers can be used (see \code{?stringr::modifiers})
and then \code{stringr::str_detect()} is used.
}
